refactor: changes App dir
