+++ frontend/src/App/App.jsx
@@ -1,13 +1,13 @@
 import './App.scss';
 import { Routes, Route } from 'react-router-dom';
-import Home from '../../Pages/Home/Home';
-import Budget from '../../Pages/Budget/Budget';
-import Statistic from '../../Pages/Statistic/Statistic';
-import Help from '../../Pages/Help/Help';
-import Profile from '../../Pages/Profile/Profile';
-import Header from '../Header/Header';
-import Footer from '../Footer/Footer';
-import EditProfile from '../../Pages/EditProfile/EditProfile';
+import Home from '../Pages/Home/Home';
+import Budget from '../Pages/Budget/Budget';
+import Statistic from '../Pages/Statistic/Statistic';
+import Help from '../Pages/Help/Help';
+import Profile from '../Pages/Profile/Profile';
+import Header from '../Components/Header/Header';
+import Footer from '../Components/Footer/Footer';
+import EditProfile from '../Pages/EditProfile/EditProfile';
 
 function App() {
   return (
